Add api-extractor for dts rollups and API reviews
This commit adds API extractor and an npm task `extract-api`. It does
the following:

* Generates a dts rollup file of this library's entire surface area.
  This is presently being saved to /typings/index-rollup.d.ts, but
  should probably go into a different place eventually.
* Adds API baselines under ./etc. Every run of `extract-api` will check
  if the public surface area changed, and if so, say something on the
  console. A git diff of ./etc/service-bus.api.ts shows pretty clearly
  what APIs have changed.

A lot of warnings are generated here. Most seem to be due to different
doc tag conventions, but a few suggest maybe some pieces need to be
exported. Check the generated API file for more details (search for
"WARNING:").

Before accepting this PR, we should at least:

* [ ] Decide if we want to enable the API Review workflow or turn it off
* [ ] Confirm the rollup is accurate and if so replace the typings folder